requirements.txt
opencv-python==4.6.0.66
streamlit==1.14.1
numpy==1.22.4
Pillow==9.2.0